+++ postcss.config.js
@@ -2,7 +2,7 @@ var tailwindcss = require('tailwindcss');
 
 module.exports = {
   plugins: [
-    require('postcss-import'),
+    // require('postcss-import'),
     require('tailwindcss/nesting'),
     tailwindcss('./tailwind.config.js'),
     require('autoprefixer')