+++ src/web/assets/videos/VideosAsset.php
@@ -10,7 +10,6 @@ namespace dukt\videos\web\assets\videos;
 use craft\web\AssetBundle;
 use craft\web\assets\cp\CpAsset;
 use craft\web\assets\vue\VueAsset;
-use dukt\videos\Plugin;
 
 class VideosAsset extends AssetBundle
 {
@@ -22,19 +21,12 @@ class VideosAsset extends AssetBundle
      */
     public function init(): void
     {
+        $this->sourcePath = __DIR__ . '/dist';
+
         $this->depends = [
             CpAsset::class,
             VueAsset::class,
         ];
-
-        if (!Plugin::getInstance()->getVideos()->useDevServer) {
-            $this->sourcePath = __DIR__ . '/dist';
-            $this->js[] = 'main.js';
-            $this->css[] = 'css/main.css';
-        } else {
-            $this->css[] = 'https://localhost:8090/css/main.css';
-            $this->js[] = 'https://localhost:8090/main.js';
-        }
 
         parent::init();
     }