+++ tailwind.config.js
@@ -12,8 +12,5 @@ module.exports = {
   },
   variants: {
     extend: {},
-  },
-  plugins: [
-    require('@tailwindcss/line-clamp'),
-  ],
+  }
 }